refactor(ui): MessageList 组件完整拆分到 message/ 目录
将 MessageList.tsx（1861 行）的所有子组件和工具函数拆分到
message/ 目录的独立文件：

message/
├── types.ts              类型定义（已有）
├── utils.ts              纯工具函数（已有）
├── useExpansionState.ts  展开/收起 hook
├── ContentMedia.tsx      媒体内容渲染
├── MessageActions.tsx    消息操作（复制/TTS）
├── UserMessageActions.tsx 用户消息操作（编辑/复制）
├── VoiceBubble.tsx       语音气泡
├── StreamingMessage.tsx  流式消息
├── ToolGroup.tsx         工具卡片组
├── UserMessageGroup.tsx  用户消息组（含编辑）
├── AgentTurn.tsx         Agent 回合渲染
└── index.ts              统一导出

MessageList.tsx 从 1861 行降到 1016 行（-845 行），
只保留主组件（虚拟滚动 + 消息分组 + 状态管理）。
