+++ frontend/src/components/MessageList.tsx
@@ -1,35 +1,19 @@
 import { useStore } from "@/store/useStore";
 import { useSearchStore } from "@/store/useSearchStore";
-import { findSearchMatches, findTextOccurrences } from "@/utils/search";
-import { HighlightText } from "./HighlightText";
+import { findSearchMatches } from "@/utils/search";
 import { SearchBar } from "./SearchBar";
 import { ScrollArea } from "./ui/scroll-area";
-import { Textarea } from "./ui/textarea";
 import {
   Loader2,
-  ChevronRight,
-  ChevronDown,
   Cpu,
-  FileText,
-  Volume2,
-  Square,
-  Copy,
-  Check,
-  Play,
-  ChevronUp,
   ShieldCheck,
   ShieldX,
-  Pencil,
-  X,
-  Paperclip,
   ArrowUp,
   ArrowDown,
   ArrowDownToLine,
 } from "lucide-react";
-import { MdPreview } from 'md-editor-rt';
 import 'md-editor-rt/lib/preview.css';
 import { open } from '@tauri-apps/plugin-dialog';
-import { ThinkingBlock } from "./ThinkingBlock";
 import { AgentPanel } from "./AgentPanel";
 import { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider } from "@/components/ui/tooltip";
 import { api, textContent, type ContentBlock } from "@/api/tauri";
@@ -39,366 +23,18 @@ import {
   fileToDataUrl,
   attachmentFromPath,
   estimatedBase64Size,
-  resolveAttachmentUrl,
   attachmentsToBase64Media,
 } from '@/utils/attachments';
 import { useVirtualizer } from "@tanstack/react-virtual";
 
-import { memo, useEffect, useMemo, useRef, useState, useCallback } from "react";
-import { useResolvedTheme } from "@/hooks/useTheme";
+import { useEffect, useMemo, useRef, useState, useCallback } from "react";
 import {
-  type MessageItem,
-  type MessageGroup,
-  formatMessageTime,
-  msgReasoning,
-  resolveAssetUrl,
-  resolveMarkdownImages,
-  extractLlmExplanation,
-  llmOutputHasToolCalls,
-  summarizeToolGroup,
-  getToolMessageMeta,
   groupMessages,
   workerContentMessages,
-  sameMessageRefs,
-  hasMessage,
   extractAgentRoles,
-  parseAgentReply,
+  UserMessageGroup,
+  AgentTurn,
 } from "./message";
-
-/** 格式化消息时间（hover 显示） */
-function MessageActions({ text, showTts }: { text: string; showTts: boolean }) {
-  const [copied, setCopied] = useState(false);
-  const [playing, setPlaying] = useState(false);
-  const [ttsLoading, setTtsLoading] = useState(false);
-
-  const handleCopy = async () => {
-    try {
-      await navigator.clipboard.writeText(text);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 2000);
-    } catch (e) {
-      console.error("复制失败:", e);
-    }
-  };
-
-  const handleTts = async () => {
-    if (playing) {
-      api.stopAudio().catch(() => {});
-      setPlaying(false);
-      return;
-    }
-
-    setTtsLoading(true);
-    try {
-      api.stopAudio().catch(() => {});
-      const result = await api.synthesizeSpeech(text);
-      setPlaying(true);
-      setTtsLoading(false);
-      await api.playAudioFile(result.file_path);
-      setPlaying(false);
-    } catch (e: any) {
-      console.error("TTS 播放失败:", e);
-      alert(`语音播放失败：${e?.message || e}`);
-      setPlaying(false);
-      setTtsLoading(false);
-    }
-  };
-
-  const btnClass = "p-1 rounded text-muted-foreground hover:text-foreground hover:bg-accent transition-colors";
-
-  return (
-    <div className="flex items-center gap-0.5 mt-1">
-      <button onClick={handleCopy} className={btnClass} title={copied ? "已复制" : "复制"}>
-        {copied ? <Check className="w-3.5 h-3.5" /> : <Copy className="w-3.5 h-3.5" />}
-      </button>
-      {showTts && (
-        <button onClick={handleTts} className={btnClass} title={playing ? "停止播放" : "朗读"}>
-          {ttsLoading ? (
-            <Loader2 className="w-3.5 h-3.5 animate-spin" />
-          ) : playing ? (
-            <Square className="w-3.5 h-3.5" />
-          ) : (
-            <Volume2 className="w-3.5 h-3.5" />
-          )}
-        </button>
-      )}
-    </div>
-  );
-}
-
-function UserMessageActions({ text, messageId, runStatus, canEdit, onStartEdit }: {
-  text: string;
-  messageId: string;
-  runStatus: string;
-  canEdit: boolean;
-  onStartEdit: (messageId: string, text: string) => void;
-}) {
-  const [copied, setCopied] = useState(false);
-  const idle = runStatus === "idle";
-
-  const handleCopy = async () => {
-    try {
-      await navigator.clipboard.writeText(text);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 2000);
-    } catch (e) {
-      console.error("复制失败:", e);
-    }
-  };
-
-  const btnClass = "p-1 rounded text-muted-foreground/50 hover:text-muted-foreground hover:bg-accent transition-colors";
-
-  return (
-    <div className="flex items-center gap-0.5 mt-1">
-      <button onClick={handleCopy} className={btnClass} title={copied ? "已复制" : "复制"}>
-        {copied ? <Check className="w-3.5 h-3.5" /> : <Copy className="w-3.5 h-3.5" />}
-      </button>
-      <button
-        onClick={() => onStartEdit(messageId, text)}
-        className={`${btnClass} ${(!idle || !canEdit) ? 'opacity-30 cursor-not-allowed' : ''}`}
-        title={!canEdit ? "已压缩消息无法编辑" : !idle ? "执行中无法编辑" : "编辑并重发"}
-        disabled={!idle || !canEdit}
-      >
-        <Pencil className="w-3.5 h-3.5" />
-      </button>
-    </div>
-  );
-}
-
-function VoiceBubble({ messageId, audioPath, duration, showText, content }: {
-  messageId: string;
-  audioPath: string;
-  duration?: number;
-  showText: boolean;
-  content: string;
-}) {
-  const [playing, setPlaying] = useState(false);
-  const { toggleVoiceText } = useStore();
-
-  const handlePlay = async () => {
-    if (playing) {
-      await api.stopAudio().catch(() => {});
-      setPlaying(false);
-      return;
-    }
-    setPlaying(true);
-    try {
-      await api.playAudioFile(audioPath);
-    } catch (e) {
-      console.error("播放语音失败:", e);
-    }
-    setPlaying(false);
-  };
-
-  return (
-    <div>
-      <button
-        className="flex items-center gap-2 text-sm hover:opacity-80 transition-opacity"
-        onClick={handlePlay}
-        title={playing ? "停止播放" : "点击播放语音"}
-      >
-        {playing ? (
-          <Square className="w-4 h-4 shrink-0" />
-        ) : (
-          <Play className="w-4 h-4 shrink-0 fill-current" />
-        )}
-        <div className="flex items-center gap-1">
-          <span className="inline-block w-16 h-[3px] rounded bg-foreground/40" />
-          <span className="text-xs text-muted-foreground">
-            {duration ? `${Math.round(duration)}″` : '语音'}
-          </span>
-        </div>
-      </button>
-      <div className="mt-1">
-        <button
-          className="text-xs text-muted-foreground hover:text-foreground transition-colors"
-          onClick={() => toggleVoiceText(messageId)}
-        >
-          {showText ? (
-            <ChevronUp className="w-3 h-3 inline mr-0.5" />
-          ) : (
-            <ChevronDown className="w-3 h-3 inline mr-0.5" />
-          )}
-          {showText ? '隐藏文字' : '显示文字'}
-        </button>
-      </div>
-      {showText && (
-        <p className="whitespace-pre-wrap break-words text-sm mt-1 pt-1 border-t border-border">
-          {content}
-        </p>
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Markdown 渲染组件
-// ---------------------------------------------------------------------------
-// 用户消息组渲染
-// ---------------------------------------------------------------------------
-
-function UserMessageGroup({ group, runStatus, nonEditableIds, voiceMessages, editingMessageId, editingContent, editingAttachments, editingTextareaRef, hasMultimodal, onStartEdit, onConfirmEdit, onCancelEdit, onSetEditingContent, onSetEditingAttachments, onAttachFiles, onEditPaste }: {
-  group: MessageGroup;
-  runStatus: string;
-  nonEditableIds: Set<string>;
-  voiceMessages: Record<string, { audioPath: string; duration?: number; showText: boolean }>;
-  editingMessageId: string | null;
-  editingContent: string;
-  editingAttachments: Attachment[];
-  editingTextareaRef: React.RefObject<HTMLTextAreaElement>;
-  hasMultimodal: boolean;
-  onStartEdit: (messageId: string, text: string) => void;
-  onConfirmEdit: () => void;
-  onCancelEdit: () => void;
-  onSetEditingContent: (v: string) => void;
-  onSetEditingAttachments: React.Dispatch<React.SetStateAction<Attachment[]>>;
-  onAttachFiles: () => void;
-  onEditPaste: (e: React.ClipboardEvent<HTMLTextAreaElement>) => void;
-}) {
-  const message = group.messages[0];
-  const voiceInfo = voiceMessages[message.id];
-  const isEditing = editingMessageId === message.id;
-  const searchQuery = useSearchStore(s => s.searchQuery);
-  const currentMessageId = useSearchStore(s => s.currentMessageId);
-  const currentMatchStart = useSearchStore(s => s.currentMatchStart);
-  const caseSensitive = useSearchStore(s => s.caseSensitive);
-
-  const renderUserText = (text: string) => {
-    if (!searchQuery) return text;
-    const occurrences = findTextOccurrences(text, searchQuery, caseSensitive);
-    if (occurrences.length === 0) return text;
-    const isCurrent = message.id === currentMessageId;
-    return <HighlightText text={text} matches={occurrences} currentMatchStart={isCurrent ? currentMatchStart : null} />;
-  };
-
-  return (
-    <div className="mt-3 first:mt-0">
-      {isEditing ? (
-        <div className="w-full">
-          {editingAttachments.length > 0 && (
-            <div className="mb-2 flex flex-wrap gap-1.5">
-              {editingAttachments.map(item => (
-                <span
-                  key={item.title + item.url.slice(0, 40)}
-                  className="inline-flex h-9 max-w-[260px] items-center gap-1.5 rounded-md border bg-muted/40 px-2 text-xs"
-                  title={item.title}
-                >
-                  {item.kind === 'image' ? (
-                    <img
-                      src={resolveAttachmentUrl(item.url)}
-                      alt={item.title}
-                      className="h-6 w-6 shrink-0 rounded object-cover"
-                    />
-                  ) : (
-                    <Paperclip className="h-3 w-3 shrink-0" />
-                  )}
-                  <span className="truncate">{item.title}</span>
-                  <button
-                    type="button"
-                    onClick={() => onSetEditingAttachments(prev => prev.filter(a => a.url !== item.url))}
-                    className="ml-1 text-muted-foreground hover:text-foreground"
-                    title="移除附件"
-                  >
-                    <X className="h-3 w-3" />
-                  </button>
-                </span>
-              ))}
-            </div>
-          )}
-          <Textarea
-            ref={editingTextareaRef}
-            value={editingContent}
-            onChange={(e) => {
-              onSetEditingContent(e.target.value);
-              const textarea = editingTextareaRef.current;
-              if (textarea) {
-                textarea.style.height = '60px';
-                textarea.style.height = Math.min(textarea.scrollHeight, 200) + 'px';
-              }
-            }}
-            onKeyDown={(e) => {
-              if (e.key === "Enter" && !e.shiftKey && !e.nativeEvent.isComposing && e.keyCode !== 229) {
-                e.preventDefault();
-                onConfirmEdit();
-              }
-              if (e.key === "Escape") {
-                onCancelEdit();
-              }
-            }}
-            onPaste={onEditPaste}
-            className="min-h-[60px] max-h-[200px] resize-none text-sm w-full"
-            autoFocus
-          />
-          <div className="flex justify-between items-center mt-1">
-            <span className="text-[10px] text-muted-foreground">Enter 发送 · Shift+Enter 换行 · Esc 取消</span>
-            <div className="flex gap-1.5">
-              {hasMultimodal && (
-                <button
-                  onClick={onAttachFiles}
-                  className="flex items-center gap-1 px-2 py-1 text-xs text-muted-foreground hover:text-foreground transition-colors"
-                  title="添加附件"
-                >
-                  <Paperclip className="w-3 h-3" />
-                </button>
-              )}
-              <button
-                onClick={onCancelEdit}
-                className="flex items-center gap-1 px-2 py-1 text-xs text-muted-foreground hover:text-foreground transition-colors"
-              >
-                <X className="w-3 h-3" />
-                取消
-              </button>
-              <button
-                onClick={onConfirmEdit}
-                className="px-2.5 py-1 text-xs bg-green-600 hover:bg-green-700 text-white rounded transition-colors"
-              >
-                发送
-              </button>
-            </div>
-          </div>
-        </div>
-      ) : (
-      <div className="flex justify-end" title={formatMessageTime(message.created_at)}>
-        <div className="max-w-[85%] rounded-2xl bg-primary/10 px-4 py-2.5 text-foreground">
-          {voiceInfo ? (
-            <VoiceBubble
-              messageId={message.id}
-              audioPath={voiceInfo.audioPath}
-              duration={voiceInfo.duration}
-              showText={voiceInfo.showText}
-              content={textContent(message)}
-            />
-          ) : (
-            <div>
-              {renderContentMedia(message)}
-              {textContent(message) && (
-                <p className="whitespace-pre-wrap break-words text-sm">
-                  {renderUserText(textContent(message))}
-                </p>
-              )}
-            </div>
-          )}
-        </div>
-      </div>
-      )}
-      {textContent(message) && !isEditing && (
-        <div className="flex justify-end">
-          <UserMessageActions
-            text={textContent(message)}
-            messageId={message.id}
-            runStatus={runStatus}
-            canEdit={!nonEditableIds.has(message.id)}
-            onStartEdit={onStartEdit}
-          />
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// 主组件
-// ---------------------------------------------------------------------------
 
 export function MessageList() {
   const messages = useStore(s => s.messages);
@@ -1378,484 +1014,3 @@ export function MessageList() {
     </div>
   );
 }
-
-// ---------------------------------------------------------------------------
-// 消息分组：User 消息单独成组，其余合并为 agent_turn
-// ---------------------------------------------------------------------------
-
-function renderContentMedia(message: MessageItem) {
-  const content = Array.isArray(message.content) ? message.content : [];
-  const mediaBlocks = content.filter((b) => b.type === 'media');
-  const legacyMedia = message.media || [];
-  const allMedia = [
-    ...mediaBlocks.map((b) => ({ kind: b.kind!, url: b.url!, title: b.title, mime_type: b.mime_type })),
-    ...legacyMedia,
-  ];
-  if (allMedia.length === 0) {
-    return null;
-  }
-
-  return (
-    <div className="space-y-2 my-2">
-      {allMedia.map((asset, index) => {
-        const src = resolveAssetUrl(asset.url);
-        if (asset.kind === "image") {
-          return (
-            <img
-              key={`${message.id}-media-${index}`}
-              src={src}
-              alt={asset.title || "生成的图片"}
-              className="max-w-full max-h-96 rounded-md cursor-pointer hover:opacity-90 transition-opacity"
-              loading="lazy"
-            />
-          );
-        }
-
-        if (asset.kind === "video") {
-          return (
-            <video
-              key={`${message.id}-media-${index}`}
-              src={src}
-              controls
-              className="max-w-full max-h-96 rounded-md"
-              preload="metadata"
-            >
-              {asset.title || "生成的视频"}
-            </video>
-          );
-        }
-
-        if (asset.kind === "audio") {
-          return (
-            <audio
-              key={`${message.id}-media-${index}`}
-              src={src}
-              controls
-              className="w-full"
-              preload="metadata"
-            >
-              {asset.title || "生成的音频"}
-            </audio>
-          );
-        }
-
-        return (
-          <a
-            key={`${message.id}-media-${index}`}
-            href={src}
-            className="text-blue-400 hover:text-blue-300 underline text-sm"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            {asset.title || asset.url}
-          </a>
-        );
-      })}
-    </div>
-  );
-}
-
-function StreamingMessage({
-  content,
-  reasoningContent,
-}: {
-  content: string;
-  reasoningContent: string;
-}) {
-  const resolvedTheme = useResolvedTheme();
-  return (
-    <div>
-      {reasoningContent && (
-        // 本组件仅在流式输出（活跃态）时渲染，故 thinking 始终默认展开以实时跟随
-        <ThinkingBlock content={reasoningContent} defaultExpanded={true} />
-      )}
-      <MdPreview modelValue={resolveMarkdownImages(content)} theme={resolvedTheme} previewTheme="github" />
-      {content.length > 0 && (
-        <span className="inline-block w-1.5 h-4 bg-primary ml-0.5 animate-pulse" />
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// 从 LLM 输出系统消息中提取解释文本
-// ---------------------------------------------------------------------------
-
-/** 统一的智能体回合渲染 — 将系统消息（事件）和 assistant 消息（回复）合并展示 */
-interface AgentTurnProps {
-  messages: MessageItem[];
-  streamingMessageId: string | null;
-  streamingContent: string;
-  streamingReasoningContent: string;
-  hasTts: boolean;
-  selectedAgentTab: string | null;
-  /** 是否为当前正在进行的轮次（输出中）：强制展开 thinking/工具/记忆，完成后随 key 重挂载自动收起 */
-  isActive?: boolean;
-}
-
-/**
- * 展开/收起状态：活跃态默认全展开，用户仍可手动收起（记录到 userCollapsed 覆盖默认）；
- * 非活跃态默认全收起，用户手动展开才进入 expanded。
- * 轮次完成（key 变化）触发组件重挂载，两个集合都会重置。
- */
-function useExpansionState(isActive: boolean) {
-  const [expanded, setExpanded] = useState<Set<string>>(new Set());
-  const [userCollapsed, setUserCollapsed] = useState<Set<string>>(new Set());
-
-  const isExpanded = (id: string) =>
-    isActive ? !userCollapsed.has(id) : expanded.has(id);
-
-  const toggle = useCallback((id: string) => {
-    if (isActive) {
-      setUserCollapsed((prev) => {
-        const next = new Set(prev);
-        next.has(id) ? next.delete(id) : next.add(id);
-        return next;
-      });
-    } else {
-      setExpanded((prev) => {
-        const next = new Set(prev);
-        next.has(id) ? next.delete(id) : next.add(id);
-        return next;
-      });
-    }
-  }, [isActive]);
-
-  return { isExpanded, toggle };
-}
-
-/** 从 agent_event 内容中提取相关 agent role */
-function AgentTurnView({
-  messages,
-  streamingMessageId,
-  streamingContent,
-  streamingReasoningContent,
-  hasTts,
-  selectedAgentTab,
-  isActive = false,
-}: AgentTurnProps) {
-  const searchQuery = useSearchStore(s => s.searchQuery);
-  const currentMessageId = useSearchStore(s => s.currentMessageId);
-  const currentMatchStart = useSearchStore(s => s.currentMatchStart);
-  const caseSensitive = useSearchStore(s => s.caseSensitive);
-  const itemExpansion = useExpansionState(isActive);
-  const toolGroupExpansion = useExpansionState(isActive);
-  const agents = useStore((state) => state.agents);
-  const resolvedTheme = useResolvedTheme();
-
-  /** 带搜索高亮渲染文本（搜索激活时使用 HighlightText，否则返回原始文本） */
-  const renderWithHighlight = (msgId: string, text: string) => {
-    if (!searchQuery) return text;
-    const occurrences = findTextOccurrences(text, searchQuery, caseSensitive);
-    if (occurrences.length === 0) return text;
-    const isCurrent = msgId === currentMessageId;
-    return <HighlightText text={text} matches={occurrences} currentMatchStart={isCurrent ? currentMatchStart : null} />;
-  };
-
-  // 将消息序列解析为渲染片段
-  type Fragment =
-    | { type: "explanation"; text: string; time?: string }
-    | { type: "thinking"; content: string; time?: string }
-    | { type: "tool_group"; key: string; brief: string; tools: MessageItem[] }
-    | { type: "user"; msg: MessageItem }
-    | { type: "assistant"; msg: MessageItem; isStreaming: boolean }
-    | { type: "error_system"; msg: MessageItem }
-    | { type: "retry_system"; msg: MessageItem }
-    | { type: "context_management"; msg: MessageItem }
-    | { type: "agent_event"; category: string; content: string; agentRoles: string[] }
-    | { type: "other_system"; msg: MessageItem };
-
-  const fragments: Fragment[] = [];
-  const shownReasonings = new Set<string>();
-  let pendingTools: MessageItem[] = [];
-
-  const flushTools = () => {
-    if (pendingTools.length === 0) return;
-    const key = pendingTools[0].id;
-    fragments.push({
-      type: "tool_group",
-      key,
-      brief: summarizeToolGroup(pendingTools),
-      tools: [...pendingTools],
-    });
-    pendingTools = [];
-  };
-
-  for (const msg of messages) {
-    if (msg.role === "user") {
-      flushTools();
-      fragments.push({ type: "user", msg });
-    } else if (msg.role === "system" && textContent(msg).startsWith("[记忆检索] 策略:")) {
-      // recall_memory 策略消息：加入 pendingTools（统一工具卡片）
-      pendingTools.push(msg);
-    } else if (msg.role === "system" && textContent(msg).startsWith("[记忆检索]")) {
-      // recall_memory 结果消息：加入 pendingTools
-      pendingTools.push(msg);
-    } else if (msg.role === "system" && textContent(msg).startsWith("LLM 输出")) {
-      const reasoning = msgReasoning(msg);
-      const explanation = extractLlmExplanation(textContent(msg));
-      if (!reasoning && !explanation && llmOutputHasToolCalls(textContent(msg))) {
-        continue;
-      }
-      flushTools();
-      if (reasoning && !shownReasonings.has(reasoning)) {
-        shownReasonings.add(reasoning);
-        fragments.push({ type: "thinking", content: reasoning, time: msg.created_at });
-      }
-      if (explanation) {
-        fragments.push({ type: "explanation", text: explanation, time: msg.created_at });
-      }
-    } else if (msg.role === "system" && (textContent(msg).includes("tool_name:") || textContent(msg).includes("exit_code") || textContent(msg).startsWith("工具执行 ["))) {
-      pendingTools.push(msg);
-    } else if (msg.role === "tool") {
-      // 所有 tool 消息（含 recall_memory、plugin_injection）统一加入 pendingTools
-      pendingTools.push(msg);
-      continue;
-    } else if (msg.role === "assistant") {
-      const isStreaming = msg.id === streamingMessageId;
-      const assistantReasoning = msgReasoning(msg);
-      const hasVisibleAssistantContent =
-        isStreaming ||
-        textContent(msg).trim().length > 0 ||
-        assistantReasoning.length > 0 ||
-        !!msg.media?.length ||
-        msg.content.some(b => b.type === "media");
-      if (!hasVisibleAssistantContent) {
-        continue;
-      }
-
-      flushTools();
-      const prevFrag = fragments[fragments.length - 1];
-      if (prevFrag?.type === "explanation" && prevFrag.text === textContent(msg).trim() && !isStreaming) {
-        fragments.pop();
-      }
-      if (!isStreaming && assistantReasoning && !shownReasonings.has(assistantReasoning)) {
-        shownReasonings.add(assistantReasoning);
-        fragments.push({ type: "thinking", content: assistantReasoning, time: msg.created_at });
-      }
-      fragments.push({ type: "assistant", msg, isStreaming });
-    } else if (msg.role === "system" && textContent(msg).startsWith("[错误]")) {
-      flushTools();
-      fragments.push({ type: "error_system", msg });
-    } else if (msg.role === "system" && textContent(msg).startsWith("[重试]")) {
-      flushTools();
-      fragments.push({ type: "retry_system", msg });
-    } else if (msg.role === "system" && textContent(msg).startsWith("[上下文管理]")) {
-      if (textContent(msg).includes("正在压缩")) {
-        continue;
-      }
-      flushTools();
-      fragments.push({ type: "context_management", msg });
-    } else if (msg.role === "system" && (
-      textContent(msg).startsWith("[Agent]")
-      || textContent(msg).startsWith("[文件锁]")
-    )) {
-      flushTools();
-      let category = "info";
-      if (textContent(msg).startsWith("[文件锁]")) category = "lock";
-      fragments.push({ type: "agent_event", category, content: textContent(msg), agentRoles: extractAgentRoles(textContent(msg), agents) });
-    } else if (msg.role === "system") {
-      flushTools();
-      fragments.push({ type: "other_system", msg });
-    }
-  }
-  flushTools();
-
-  const mergedFragments: Fragment[] = [];
-  for (const frag of fragments) {
-    const previous = mergedFragments[mergedFragments.length - 1];
-    if (frag.type === "tool_group" && previous?.type === "tool_group") {
-      previous.tools.push(...frag.tools);
-      previous.brief = summarizeToolGroup(previous.tools);
-      continue;
-    }
-    mergedFragments.push(frag);
-  }
-
-  /** 渲染工具条目 */
-  const renderToolItem = (tool: MessageItem) => {
-    const meta = getToolMessageMeta(tool);
-    const expanded = itemExpansion.isExpanded(tool.id);
-    return (
-      <div key={tool.id} title={formatMessageTime(tool.created_at)}>
-        <button
-          className="w-full flex items-center gap-2 px-2 py-0.5 rounded text-xs text-muted-foreground hover:bg-muted/50 transition-colors text-left"
-          onClick={() => itemExpansion.toggle(tool.id)}
-        >
-          {expanded ? <ChevronDown className="w-3 h-3 shrink-0" /> : <ChevronRight className="w-3 h-3 shrink-0" />}
-          <meta.icon className="w-3 h-3 shrink-0" />
-          <span className="font-medium">{meta.label}</span>
-          {!expanded && <span className="truncate opacity-60">{meta.summary}</span>}
-        </button>
-        {expanded && (
-          <div className="ml-5 mt-0.5 px-3 py-2 rounded-md bg-muted/30 border border-border/50">
-            <pre className="text-xs text-muted-foreground whitespace-pre-wrap break-words font-mono leading-relaxed">{textContent(tool)}</pre>
-          </div>
-        )}
-      </div>
-    );
-  };
-
-  return (
-    <div className="space-y-1.5">
-      {mergedFragments.map((frag, i) => {
-        if (selectedAgentTab && frag.type !== "agent_event") {
-          return null;
-        }
-        if (selectedAgentTab && frag.type === "agent_event" && frag.agentRoles.length > 0 && !frag.agentRoles.includes(selectedAgentTab)) {
-          return null;
-        }
-        if (frag.type === "thinking") {
-          return (
-            <div key={`think-${i}`} title={formatMessageTime(frag.time)}>
-              <ThinkingBlock content={frag.content} defaultExpanded={isActive} />
-            </div>
-          );
-        }
-        if (frag.type === "explanation") {
-          return (
-            <p key={`expl-${i}`} className="text-sm text-muted-foreground leading-relaxed whitespace-pre-wrap break-words" title={formatMessageTime(frag.time)}>
-              {frag.text}
-            </p>
-          );
-        }
-        if (frag.type === "tool_group") {
-          const collapsed = !toolGroupExpansion.isExpanded(frag.key);
-          const groupTime = frag.tools.length > 0 ? formatMessageTime(frag.tools[0].created_at) : "";
-          return (
-            <div key={`tools-${frag.key}`} title={groupTime}>
-              <button
-                className="flex items-center gap-2 px-2 py-0.5 text-xs text-muted-foreground hover:bg-muted/50 rounded transition-colors"
-                onClick={() => toolGroupExpansion.toggle(frag.key)}
-              >
-                {collapsed ? <ChevronRight className="w-3 h-3 shrink-0" /> : <ChevronDown className="w-3 h-3 shrink-0" />}
-                <Cpu className="w-3 h-3 shrink-0" />
-                <span>{frag.brief}</span>
-              </button>
-              {!collapsed && <div className="ml-4 space-y-0">{frag.tools.map(t => renderToolItem(t))}</div>}
-            </div>
-          );
-        }
-        if (frag.type === "user") {
-          return (
-            <div key={frag.msg.id} className="flex justify-end" title={formatMessageTime(frag.msg.created_at)}>
-              <div className="max-w-[85%] rounded-2xl bg-primary/10 px-4 py-2.5 text-sm text-foreground whitespace-pre-wrap break-words">
-                {textContent(frag.msg)}
-              </div>
-            </div>
-          );
-        }
-        if (frag.type === "assistant") {
-          const { msg, isStreaming } = frag;
-          const agentReply = !isStreaming ? parseAgentReply(textContent(msg)) : null;
-          if (agentReply) {
-            return (
-              <div key={msg.id} className="text-foreground" title={formatMessageTime(msg.created_at)}>
-                <div className="inline-flex items-center gap-1.5 rounded-full border border-green-500/30 bg-green-500/10 px-2 py-0.5 text-xs font-medium text-green-700 dark:text-green-300 mb-1">
-                  <span className="h-1.5 w-1.5 rounded-full bg-green-500" />
-                  {agentReply.label}
-                </div>
-                <div className="border-l-2 border-green-500/50 pl-3">
-                  {agentReply.body ? (
-                    searchQuery && findTextOccurrences(textContent(msg), searchQuery, caseSensitive).length > 0
-                      ? <div className="text-sm whitespace-pre-wrap break-words">{renderWithHighlight(msg.id, agentReply.body)}</div>
-                      : <MdPreview modelValue={resolveMarkdownImages(agentReply.body)} theme={resolvedTheme} previewTheme="github" />
-                  ) : null}
-                </div>
-                {agentReply.body && <MessageActions text={agentReply.body} showTts={hasTts} />}
-              </div>
-            );
-          }
-          return (
-            <div key={msg.id} className="text-foreground" title={formatMessageTime(msg.created_at)}>
-              {isStreaming ? (
-                <StreamingMessage
-                  content={streamingContent}
-                  reasoningContent={streamingReasoningContent}
-                />
-              ) : textContent(msg) || (msg.media && msg.media.length > 0) || msg.content.some(b => b.type === "media") ? (
-                <div>
-                  {renderContentMedia(msg)}
-                  {searchQuery && findTextOccurrences(textContent(msg), searchQuery, caseSensitive).length > 0
-                    ? <div className="text-sm whitespace-pre-wrap break-words">{renderWithHighlight(msg.id, textContent(msg))}</div>
-                    : <MdPreview modelValue={resolveMarkdownImages(textContent(msg))} theme={resolvedTheme} previewTheme="github" />
-                  }
-                </div>
-              ) : null}
-              {!isStreaming && msg.content && <MessageActions text={textContent(msg)} showTts={hasTts} />}
-            </div>
-          );
-        }
-        if (frag.type === "error_system") {
-          return (
-            <div key={frag.msg.id} className="text-sm text-destructive bg-destructive/10 rounded-md px-3 py-2 my-1">
-              {textContent(frag.msg).replace("[错误] ", "")}
-            </div>
-          );
-        }
-        if (frag.type === "retry_system") {
-          return (
-            <div key={frag.msg.id} className="text-xs text-yellow-600 dark:text-yellow-400 bg-yellow-500/10 rounded-md px-3 py-1.5 my-0.5">
-              {textContent(frag.msg).replace("[重试] ", "")}
-            </div>
-          );
-        }
-        if (frag.type === "context_management") {
-          const text = textContent(frag.msg).replace("[上下文管理] ", "");
-          return (
-            <div
-              key={frag.msg.id}
-              className="inline-flex max-w-full items-center gap-2 rounded-md border border-border/70 bg-muted/30 px-2.5 py-1 text-xs text-muted-foreground"
-              title={formatMessageTime(frag.msg.created_at)}
-            >
-              <FileText className="h-3.5 w-3.5 shrink-0" />
-              <span className="truncate">{text}</span>
-            </div>
-          );
-        }
-        if (frag.type === "agent_event") {
-          const colorMap: Record<string, string> = {
-            lock: "border-yellow-500/30 bg-yellow-500/5",
-            info: "border-border bg-muted/30",
-          };
-          return (
-            <div key={`agent-${i}`} className={`text-xs text-muted-foreground border rounded px-2 py-1 my-0.5 ${colorMap[frag.category] || colorMap.info}`}>
-              {frag.content}
-            </div>
-          );
-        }
-        if (frag.type === "other_system") {
-          return (
-            <p key={frag.msg.id} className="text-xs text-muted-foreground">
-              {textContent(frag.msg).split("\n")[0]}
-            </p>
-          );
-        }
-        return null;
-      })}
-    </div>
-  );
-}
-
-const AgentTurn = memo(AgentTurnView, (prev, next) => {
-  if (
-    prev.hasTts !== next.hasTts
-    || !sameMessageRefs(prev.messages, next.messages)
-    || prev.selectedAgentTab !== next.selectedAgentTab
-    || prev.isActive !== next.isActive
-  ) {
-    return false;
-  }
-
-  const touchesStreamingMessage =
-    hasMessage(prev.messages, prev.streamingMessageId)
-    || hasMessage(prev.messages, next.streamingMessageId);
-  if (!touchesStreamingMessage) {
-    return true;
-  }
-
-  return prev.streamingMessageId === next.streamingMessageId
-    && prev.streamingContent === next.streamingContent
-    && prev.streamingReasoningContent === next.streamingReasoningContent;
-});
-
