+++ frontend/src/components/message/index.ts
@@ -1,2 +1,11 @@
 export * from "./types";
 export * from "./utils";
+export * from "./useExpansionState";
+export * from "./ContentMedia";
+export * from "./MessageActions";
+export * from "./UserMessageActions";
+export * from "./VoiceBubble";
+export * from "./StreamingMessage";
+export * from "./ToolGroup";
+export * from "./UserMessageGroup";
+export * from "./AgentTurn";
